Uses managed identity for API Management access
Switches from using a service principal to using a managed identity when
the hub attempts to retrieve information from API Management. A managed
identity is created and given permissions on the API Management service.
Then a federated identity credential is created to map the managed
identity onto a service account, which is the service account used by
the jupyterhub hub pod. This service account is annotated with the
managed identity client ID, and the hub pod is labelled to enable azure
workload identity.